+++ src/InstanceBuilder.ts
@@ -9,9 +9,11 @@ import isObject from './isObject';
 export type Constructor<T> = new (...args: any[]) => T;
 
 export default class InstanceBuilder<T> extends Builder<T | null> {
+  loose: boolean = false;
+
   refClass: Constructor<T> | null = null;
 
-  constructor(refClass: Constructor<T> | null = null) {
+  constructor(refClass: Constructor<T> | null = null, loose: boolean = false) {
     super('instance', null);
 
     // Nullable by default
@@ -22,6 +24,7 @@ export default class InstanceBuilder<T> extends Builder<T | null> {
         this.invariant(typeof refClass === 'function', 'A class reference is required.');
       }
 
+      this.loose = loose;
       this.refClass = refClass;
       this.addCheck(this.checkInstance, refClass);
     }
@@ -31,7 +34,9 @@ export default class InstanceBuilder<T> extends Builder<T | null> {
     if (__DEV__) {
       if (refClass) {
         this.invariant(
-          typeof refClass === 'function' && value instanceof refClass,
+          typeof refClass === 'function' &&
+            (value instanceof refClass ||
+              (this.loose && isObject(value) && value.constructor.name === refClass.name)),
           `Must be an instance of "${this.typeAlias()}".`,
           path,
         );
@@ -55,8 +60,11 @@ export default class InstanceBuilder<T> extends Builder<T | null> {
   }
 }
 
-export function instance<T = Function>(refClass: Constructor<T> | null = null) /* infer */ {
-  return new InstanceBuilder<T>(refClass);
+export function instance<T = Function>(
+  refClass: Constructor<T> | null = null,
+  loose?: boolean,
+) /* infer */ {
+  return new InstanceBuilder<T>(refClass, loose);
 }
 
 export function regex() /* infer */ {