+++ tests/ShapeBuilder.test.ts
@@ -62,9 +62,9 @@ describe('shape()', () => {
       expect(() => {
         builder.runChecks(
           'key',
+          // @ts-ignore Allow invalid type
           {
             foo: 'foo',
-            // @ts-ignore Allow invalid type
             bar: 'bar',
             baz: true,
           },
@@ -77,8 +77,8 @@ describe('shape()', () => {
       expect(() => {
         builder.runChecks(
           'key',
+          // @ts-ignore Allow invalid type
           {
-            // @ts-ignore Allow invalid type
             foo: 123,
           },
           {},